refactor(models/database): adapt to new database schema
