--- crates/models/database/src/user_to_collection.rs
@@ -1,46 +1,0 @@
-//! `SeaORM` Entity. Generated by sea-orm-codegen 0.12.15
-
-use sea_orm::entity::prelude::*;
-
-#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq)]
-#[sea_orm(table_name = "user_to_collection")]
-pub struct Model {
-    #[sea_orm(primary_key, auto_increment = false)]
-    pub collection_id: String,
-    #[sea_orm(primary_key, auto_increment = false)]
-    pub user_id: String,
-}
-
-#[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
-pub enum Relation {
-    #[sea_orm(
-        belongs_to = "super::collection::Entity",
-        from = "Column::CollectionId",
-        to = "super::collection::Column::Id",
-        on_update = "Cascade",
-        on_delete = "Cascade"
-    )]
-    Collection,
-    #[sea_orm(
-        belongs_to = "super::user::Entity",
-        from = "Column::UserId",
-        to = "super::user::Column::Id",
-        on_update = "Cascade",
-        on_delete = "Cascade"
-    )]
-    User,
-}
-
-impl Related<super::collection::Entity> for Entity {
-    fn to() -> RelationDef {
-        Relation::Collection.def()
-    }
-}
-
-impl Related<super::user::Entity> for Entity {
-    fn to() -> RelationDef {
-        Relation::User.def()
-    }
-}
-
-impl ActiveModelBehavior for ActiveModel {}
